requirements.txt
python-telegram-bot==22.1
jdatetime==4.1.1 
Pillow>=9.0.0
python-bidi>=0.4.2 
# httpx به عنوان وابستگی python-telegram-bot نصب می‌شود و معمولاً نیازی به ذکر جداگانه ندارد،
# اما اگر به طور مستقیم از آن استفاده می‌کنید یا می‌خواهید نسخه‌ی خاصی را تضمین کنید، می‌توانید اضافه کنید.
# httpx==0.28.1 
openai>=1.13.0 